// backend/src/mutual-fund/schemas/sip.schema.ts
import { z } from "zod";
import { FundType } from "@prisma/client";

export const createSipSchema = z.object({
  body: z.object({
    pin: z.coerce.number().int().positive().max(9999).optional(),
    amount: z.coerce.number().positive().max(10000000),
    sipDate: z.coerce.number().int(),
    schemeCode: z.coerce.number().int().positive(),
    fundName: z.string().trim(),
    fundShortName: z.string().trim(),
    fundCategory: z.string().trim(),
    fundHouseDomain: z.string().trim(),
    fundType: z.enum(FundType),
  }),
});

export const editSipSchema = z.object({
  params: z.object({
    sipId: z.uuid(),
  }),
  body: z
    .object({
      amount: z.coerce.number().positive().nullable().optional(),
      sipDate: z.coerce.number().int().nullable().optional(),
    })
    .refine((data) => data.amount != null || data.sipDate != null, {
      message: "amount or sipDate one is required",
      path: ["amount"],
    }),
});

export const stepUpSipSchema = z.object({
  params: z.object({
    sipId: z.uuid(),
  }),
  body: z
    .object({
      intervalInMonths: z.coerce.number().int().positive(),
      amount: z.coerce.number().positive().nullable().optional(),
      percentage: z.coerce.number().positive().max(100).nullable().optional(),
    })
    .refine((data) => data.amount != null || data.percentage != null, {
      message: "amount or percentage one is required",
      path: ["amount"],
    }),
});

export const sipIdParamSchema = z.object({
  params: z.object({
    sipId: z.uuid(),
  }),
});

// ========== Type exports ==========

export type CreateSipSchema = z.infer<typeof createSipSchema.shape.body>;
export type EditSipSchema = z.infer<typeof editSipSchema.shape.body>;
export type StepUpSipSchema = z.infer<typeof stepUpSipSchema.shape.body>;